Moved preable over to using the Visitor trait
