+++ src/lib.rs
@@ -96,7 +96,7 @@ mod document;
 mod section;
 mod lists;
 mod equations;
-mod visitor;
+pub mod visitor;
 
 pub use errors::*;
 pub use document::{Document, DocumentClass, Element, Preamble};
@@ -104,7 +104,7 @@ pub use paragraph::{Paragraph, ParagraphElement};
 pub use section::Section;
 pub use lists::{List, ListKind};
 pub use equations::{Align, Equation};
-pub use visitor::Visitor;
+pub use visitor::print;
 
 use std::io::Write;
