+++ src/paragraph.rs
@@ -92,63 +92,3 @@ impl Renderable for ParagraphElement {
         Ok(())
     }
 }
-
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use self::ParagraphElement::*;
-
-    #[test]
-    fn create_simple_paragraph() {
-        let should_be = "Hello World";
-        let mut para = Paragraph::new();
-        para.push_text("Hello World");
-
-        let mut rendered = vec![];
-        para.render(&mut rendered).unwrap();
-
-        assert_eq!(String::from_utf8(rendered).unwrap(), should_be);
-    }
-
-    #[test]
-    fn paragraph_with_bold_text() {
-        let should_be = r"Hello \textbf{World}";
-        let mut para = Paragraph::new();
-        para.push_text("Hello ");
-        para.push(Bold(Box::new(Plain("World".to_string()))));
-
-        let mut rendered = vec![];
-        para.render(&mut rendered).unwrap();
-
-        assert_eq!(String::from_utf8(rendered).unwrap(), should_be);
-    }
-
-    #[test]
-    fn paragraph_with_italic_text() {
-        let should_be = r"Hello \textit{World}";
-        let mut para = Paragraph::new();
-        para.push_text("Hello ");
-        para.push(Italic(Box::new(Plain("World".to_string()))));
-
-        let mut rendered = vec![];
-        para.render(&mut rendered).unwrap();
-
-        assert_eq!(String::from_utf8(rendered).unwrap(), should_be);
-    }
-
-    #[test]
-    fn inline_code() {
-        let should_be = r"Hello $\lambda$ World!";
-
-        let mut para = Paragraph::new();
-        para.push_text("Hello ")
-            .push(InlineCode(r"\lambda".to_string()))
-            .push_text(" World!");
-
-        let mut rendered = vec![];
-        para.render(&mut rendered).unwrap();
-
-        assert_eq!(String::from_utf8(rendered).unwrap(), should_be);
-    }
-}
